requirements.txt
google-auth==2.16.1
google-cloud-compute==1.10.1